+++ src/REvoDesign/REvoDesign.py
@@ -50,6 +50,7 @@ from REvoDesign.shortcuts.shortcuts_on_menu import (
     menu_relax_w_ca_constraints, menu_resi_renumber, menu_rosettaligand,
     menu_sdf2rosetta_params, menu_smiles_conformer_batch,
     menu_smiles_conformer_single, menu_thermompnn)
+from REvoDesign.shortcuts.tools.openmm_utils import OpenmmSetupServerControl
 from REvoDesign.structure import PocketSearcher, SurfaceFinder
 from REvoDesign.tools.customized_widgets import (WorkerThread, decide,
                                                  getExistingDirectory,
@@ -380,7 +381,7 @@ class REvoDesignPlugin(QtWidgets.QWidget):
         )
 
         MenuActionServerMonitor(ServerControl, self.bus.ui.actionStartEditor, self.bus.ui.actionStopEditor)
-
+        MenuActionServerMonitor(OpenmmSetupServerControl, self.bus.ui.actionStart_SetupOpenMM, self.bus.ui.actionStop_SetupOpenMM)
         if self.teamwork_enabled:
             self.ws_server = REvoDesignWebSocketServer()
             self.ws_client = REvoDesignWebSocketClient()