+++ src/REvoDesign/basic/__init__.py
@@ -7,7 +7,7 @@ from .group_registries import GroupRegistryItem
 from .menu_item import MenuCollection, MenuItem
 from .mutate_runner import MutateRunnerAbstract
 from .param_toggle import ParamChangeRegister, ParamChangeRegistryItem
-from .server_monitor import MenuActionServerMonitor
+from .server_monitor import MenuActionServerMonitor,ServerControlAbstract
 
 __all__ = [
     "SingletonAbstract",
@@ -23,4 +23,5 @@ __all__ = [
     'MenuItem',
     'MenuCollection',
     'MenuActionServerMonitor',
+    'ServerControlAbstract'
 ]